LY-84689 203199 V779. Unreachable code detected. It is possible that an error is present.
emit SignalFocused(this); was added in LY 1.9 and it appears as if the programmer wanted this signal to be executed on exiting the Init(...) function. However, due to the function returning before this point, the code will never get executed. By introducing a local boolean parameter it is possibly to maintain the same flow of execution whilst also emitting the signal as per the perceived design.