LY-84680 203190, V564. The '&' or '|' operator is applied to bool type value. You've probably forgotten to include parentheses or intended to use the '&&' or '||' operator.
 This error actually demonstrates the canonical example as shown in the PVS documentation: https://www.viva64.com/en/w/v564/print/

Here we have an obvious error. It is the "!psContext->flags" subexpression that will be calculated at first and we will get either true of false. But it does not matter: whether we execute "true & HLSLCC_FLAG_AVOID_TEMP_REGISTER_ALIASING" operation or "false & HLSLCC_FLAG_AVOID_TEMP_REGISTER_ALIASING" operation, the result will be the same. The condition in this sample is always false.