LY-84683 203193, V564. The '&' or '|' operator is applied to bool type value. You've probably forgotten to include parentheses or intended to use the '&&' or '||' operator.
This line was a bit horrible and is a nice example of people trying to cram way too much onto one line.

First of all, walpha is a bool so the (walpha ? true : false) sub-expression is pointless and just lends to confusing the meaning of the exression.
Second, to address the actual error code, it is clear that the || operator should be used instead of the | operator such that if autopreset is true then optimizable reamains true, but if it is false then eAlphaContent will be evaluated.
Third, we can just initialise optimizable in place and make optimizable const, it is only used in the fprintf below anyway. It's tempting to change the fprintf to make it nicer to read, but, we don't know if there are any tools which may or may not depend on this format.