LY-84682 203192, V564. The '&' or '|' operator is applied to bool type value. You've probably forgotten to include parentheses or intended to use the '&&' or '||' operator.
This appears to be another case of getting the operator precedence wrong; !pAnimNode->GetParamFlags(paramType) will be applied first resulting in a true or false value which will have the & operator applied.
It appears that the intention is actually to check if eSupportedParamFlags_MultipleTracks flag is set: (pAnimNode->GetParamFlags(paramType) & IAnimNode::eSupportedParamFlags_MultipleTracks) and if it is, then we will continue past the code to AddButtonToTracksToolBar. By adding parenthesis we can ensure this actually happens.