LY-84681 203191, V564. The '&' or '|' operator is applied to bool type value. You've probably forgotten to include parentheses or intended to use the '&&' or '||' operator.
This is another simple case of using the | operator instead of the || operator.
From examining the code it fairly clear that the intention is that if the FOB_DYNAMIC_OBJECT flag is set, then only set stencilRef if CV_r_deferredDecalsOnDynamicObjects is non-zero. Previously, the code would never set stencilRef if the FOB_DYNAMIC_OBJECT flag is set.