LY-84679 203189 PVS Article - Ownerless token 'false'. fillspacetool.cpp 191
The analyzer has detected a potential error: an extra lexeme in the code. Such "lost" lexemes most often occur in the code when the key word return is missing.
In the case of the FillSpaceTool, if the validEdgeList is empty, the function should be returning false as there is no work to do.