LY-84675 203185, V561. It's probably better to assign value to 'foo' variable than to declare it anew.
This error code has pointed out incorrect use the curveName variable. In the existing code, a preset would be added to the library without the name correctly set. The problem is that there is a second declaration of curveName in a for loop. This curveName will be assigned to in the for loop scope, then will go out of scope before being added to the preset library. Seeing as we already have the name of the preset available in the scope of AddPresetToLibrary, let's just use that instead.