LY-84663 203173, LY-84662 203172, V557. Array overrun is possible.
The variable stateID could be used to index m_animationStates when it has a value of m_animationStates.size() causing an out of bound array access error.

I believe that this file/system is deprecated but it's worth fixing in the interim while it is in the codebase. It is feasible (though unlikely!) that someone may use this as reference and copy/paste an error into new code.