LY-84659 203169, V557. Array overrun is possible.
The variable index could be used to index m_m_seats when it has a value of m_seats.size() causing an out of bound array access error.

I believe that this file/system is deprecated but it's worth fixing in the interim while it is in the codebase. It is feasible (though unlikely!) that someone may use this as reference and copy/paste an error into new code.