The first argument of 'memcmp' function is equal to the second argument
The condition compares the sizes of two vectors: if they are equal, the else branch is executed, where the values of the vectors' first elements are compared using the memcmp() function. But the problem is that the first and second arguments of this function are the same. The array elements are accessed in quite a complicated way using the indexes a and b - it must be one of them that was mistyped.