LY-84652 203144, Perhaps the '?:' operator works in a different way than it was expected. The '?:' operator has a lower priority than the '&&' operator.
This is a horrible little bit of a logic to cram onto one line! I have moved the implementation into the source file and broken down the logic to make it much easier for the next person to read and understand.

As far as I can tell from the code and from the documentation of the shape tool, the intention is that if polygons are picked and the shift modifier is enabled, the shapes should be separated. The current logic is obtuse and confusing so please check my reasoning!