LY-84651 203143, V502, Perhaps the '?:' operator works in a different way than it was expected. The '?:' operator has a lower priority than the '+' operator.
This looks like the analyser got it spot on; it appears that the programmer intends to increase the indent size by 24 if the widget is not collapsible, and by 0 if it is. The problem is that they have messed up the operator precedence and will be performing the ternary operation on ((groupDepth * 24) + (!m_collapsible)) when the intention was to perform it on (!m_collapsible). Adding parenthesis sorts this bug in indentation out.