LY-84650 203142, V502 Perhaps the '?:' operator works in a different way than it was expected. The '?:' operator has a lower priority than the '-' operator.
This is a deprecated system, but to avoid any future copy and paste errors while the file still exists, we should fix this. This is another case of "should have used parenthesis". The code in question will not work as intended due to order of evaluation of the expression on the left of the ternary operator.