LY-84640 203130, There are identical sub-expressions '(TriMiddle - RMWPosition)' to the left and to the right of the '|' operator.
I have just been looking at this one and it had me confused at first, but the code is actually correct. The `|` operator is overloaded to perform the dot product on two Vec3s. If the vectors are parallel then the dot product just results in the ordinary multiplication of their lengths. If the vectors were normalised then this would just result in 1, but as they are not we get the squared distance. It's not clear, but it is correct! Let's add a comment to help the next person to look at this out!