LY-84626 203106, LY-84625 203105, V501. There are identical sub-expressions to the left and to the right of the 'foo' operator.
Fix for a few typos in this file where `pRT_ALD_1` and `pRT_ALD_0` have been entered instead of  `pRT_RGB_1` and `pRT_RGB_0`. These are the texture tracing targets and by looking at the code the conditionals are protecting, it is clear to see that these are simply null checking them.

There were also a couple of straightforward duplicated conditionals.