LY-84646 203136, V501 There are identical sub-expressions 'm_listRect.contains(event->pos())' to the left and to the right of the '||' operator.
Copy/paste error or possibly autocorrect error. Mouse button down events were not being handled correctly in the details window for the ai debugger. This is actually an easy one to test by breakpointing the affected function and clicking around the ai debugger. It is clear that the code is intended to handle clicks in the details rect on the right of the window but the old code does not handle them correctly. This is in a system which I believe will be deprecated/removed but it is worth fixing this whilst the tool is still available.