LY-84639 203129, There are identical sub-expressions 'geom_colltype_solid' to the left and to the right of the '|' operator.
Removing duplicated geom_colltype_solid.

The intention of the code highlighted by the analyzer is to add flags to an attachment so that it physically interacts with collisions. One of the flags added is a duplicate so it is worth asking whether this is actually just a duplicate and should be removed, or, if the programmer intended to type something different and got the wrong autocorrect. We can examine the geom_flags enum and see what they may have intended. On examination we can see that the only relevant flags are the ones already applied and that we should remove the duplicate.