LY-84630 203111, LY-84629 203110, There are identical sub-expressions 'm_joints[i].limits[1][j]' to the left and to the right of the '-' operator.
A typo has been copied into two places in this file. The result of this is that when stepping the CArticulatedEntity simulation step, if a joint is being forced into a position outside of it's limits, the limits of the joint will not be updated as intended. The intention of the code is:

- For each dimension
- If rotation is not locked
- check if the rotation of the joint is less than the minimum limit allowed: m_joints[i].limits[0][j] - m_joints[i].qext[j]
- check if the rotation of the joint is above the maximum limit allowed: m_joints[i].qext[j] - m_joints[i].limits[1][j]
- check if the maximum limit is actually greater than the minimum limit: m_joints[i].limits[1][j] - m_joints[i].limits[0][j]
-- This is the crucial part:
--- Usually, if max > min, if Min = 0, Max = 90, then if qExt is outside the range 0-90, then the limits should be adjusted.
--- However, if min > max then we actually invert the checks, so if Min = 90, Max = 0, then qExt is outside the limits if it is within the range 0-90
- adjust the limits if rotation of the joint is outside the limitations

The final check which checks the min vs max limit and sees if they are inverted, had a typo which meant it would always evaluate to false and the limits would not be updated regardless of whether qExt was beyond it's limits.