V773. The function was exited without releasing the pointer/handle. A memory/resource leak is possible.
Typo/copy and paste error lead to a memory leak which was actually also a functionality bug too. It appears that pGroupRec should be added to the UI structure but instead pRootGroupRec was added for a second time.
I have tested this out in the UI Animation Editor and can't actually see what it does so I'm hoping someone more familiar with this system will be able to identify the change in behaviour!