V778. Two similar code fragments were found. Perhaps, this is a typo and 'X' variable should be used instead of 'Y'.
The intention of this code is to look for a key value pair in a script table and if it exists, to load in the appropriate database. In this instance, if  the key is "SoundDatabase", the code would work, however, for "fileSoundDatabase", the sound database would never be loaded in.