V568 It's odd that 'sizeof()' operator evaluates the size of a pointer to a class, but not the size of the 'vol' class object.
The argument of the sizeof() operator is a pointer to a class. In most cases this shows that the programmer forgot to dereference the pointer. In this case it is clear that vol should be dereferenced so that we are returning the size of the object, not the pointer, for the MemStats.