V555 The expression 'bufferSize - written > 0' will work as 'bufferSize != written'.
The intention of this code appears to be, if bufferSize is greater than written, we add a newline plus 0 to the buffer and return.
The problem is that as the expression is currently, it will always be true unless `bufferSize == written`. This means that if `written > bufferSize`, the code will write past the ends of the buffer.
We just want to check if `bufferSize > written` to avoid the change of unsigned overflow and off the end memory access.