Bugfix: Freeze when creating slices
This modification fixes a freeze that can happen when requesting slice instanciation while OnAsset ready is being dispatched.

If the last item in teh queue was being Dispatched and this triggered new slice creation requests, then iter was no longer pointing to the end (it was pointing to the old end) and this was causing a freeze while spinning in the loop and progressing the iterator into unknown memory regions.