Clear editor cloud data when component is deactivated
When activating the particle data is regenerated. If you go through many cycles
of activate/deactivate in the editor this causes the particle data to have a
very large amount of duplicates, which increases activation time for the cloud
component and stalls the main thread in editor, potentially for a very long
time.