Update NetBindingSystemImpl.cpp
When a client is deleting netbound entities (for which he is not the authority), this leaves the NetBinding system with pointers to invalid entities in the m_boundEntities. Causing either crashes or deletion of random entities when the authority actually signals destruction of the netreplicated slice.

I'm working around this by storing entity Ids rather than storing pointers.