Assert fix when setting scrollable entity + avoiding LHS
Switching scrollable entity leads to asserts in Bus connection code due to tha fact that previous handler was not disconnected. So this is a pretty trivial fix.

I also changed the folowing calls from using the m_scrollableEntity that was just assigned to using the entity ip passed as parameter to avoid load-hit-stores.