Don't serialize PhysX collider scale values
The scale values come from the entity parent transform, so when changing the scale anywhere in the parent hierarchy of a physx collider, the editor thinks thinks that the collider has changed, because this changes its serialized representation.

It isn't necessary to store the scale for runtime correctness, because all of the runtime colliders invariably overwrite the scale of their shapes. It isn't clear if the physics shape configuration scale should ever be serialized by PhysX